Moved Analytics into app/services
